// src/types/HttpCode.ts
export enum HTTP_CODE {
    OK = 200,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401, // không được phép
    PAYMENT_REQUIRED = 402,
    FORBIDDEN = 403, // bị cấm
    NOT_FOUND = 404,
    METHOD_NOT_ALLOWED = 405,
    REQUEST_TIMEOUT = 408,
    INTERNAL_SERVER_ERROR = 500,
    BAD_GATEWAY = 502 
}